+++ src/analysis/PartingPerimeterQuery.m
@@ -7,8 +7,8 @@ classdef PartingPerimeterQuery < handle
     end
     
     properties ( SetAccess = private, Dependent )
-        count(1,1) uint64
-        label_array(:,:,:) uint64
+        count(1,1) uint32
+        label_array(:,:,:) uint32
         binary_array(:,:,:) logical
     end
     
@@ -37,11 +37,11 @@ classdef PartingPerimeterQuery < handle
         end
         
         function value = get.count( obj )
-            value = obj.cc.NumObjects;
+            value = uint32( obj.cc.NumObjects );
         end
         
         function value = get.label_array( obj )
-            value = labelmatrix( obj.cc );
+            value = uint32( labelmatrix( obj.cc ) );
         end
         
         function value = get.binary_array( obj )
