+++ src/analysis/ProjectedPerimeterQuery.m
@@ -6,8 +6,8 @@ classdef ProjectedPerimeterQuery < handle
     end
     
     properties ( SetAccess = private, Dependent )
-        count(1,1) uint64
-        label_array(:,:) uint64
+        count(1,1) uint32
+        label_array(:,:) uint32
         binary_array(:,:) logical
     end
     
@@ -22,18 +22,18 @@ classdef ProjectedPerimeterQuery < handle
             perimeter = bwperim( projected );
             length = sum( perimeter, 'all' );
             cc = bwconncomp( perimeter );
-            cc.NumObjects = uint64( cc.NumObjects );
+            cc.NumObjects = uint32( cc.NumObjects );
             obj.cc = cc;
             obj.area = area;
             obj.length = length;
         end
         
         function value = get.count( obj )
-            value = obj.cc.NumObjects;
+            value = uint32( obj.cc.NumObjects );
         end
         
         function value = get.label_array( obj )
-            value = labelmatrix( obj.cc );
+            value = uint32( labelmatrix( obj.cc ) );
         end
         
         function value = get.binary_array( obj )
