+++ src/analysis/ThinSectionQuery.m
@@ -5,8 +5,8 @@ classdef ThinSections < handle
     % regions of castings based on some thin wall threshold.
     
     properties ( SetAccess = private )
-        count(1,1) uint64
-        label_array(:,:,:) uint64
+        count(1,1) uint32
+        label_array(:,:,:) uint32
     end
     
     methods
@@ -57,11 +57,11 @@ classdef ThinSections < handle
         end
         
         function count = get.count( obj )
-            count = uint64( obj.cc.NumObjects );
+            count = uint32( obj.cc.NumObjects );
         end
         
         function value = get.label_array( obj )
-            value = uint64( labelmatrix( obj.cc ) );
+            value = uint32( labelmatrix( obj.cc ) );
         end
     end
